+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Development
 
+* 2026-02-02 -- 0.0.4 Livewire table and jobs template
+  * Added Livewire table with add/delete modal windows
+  * Added sync/delete job templates
+  * Updated README.md
+
 * 2026-02-02 -- 0.0.3 Models and Channel Controller
   * Added YoutubeChannel and YoutubeVideo models
   * Added YoutubeChannelController for index